+++ src/root.zig
@@ -2,7 +2,7 @@
 
 const std = @import("std");
 
-pub const version = "0.3.6";
+pub const version = "0.3.7";
 
 pub const bech32 = @import("bech32.zig");
 pub const nip19 = @import("nip19.zig");